See description for details.... ;)
Removed the mailer from the UserCounter and seperated it, so you can have a central database server and send the mails from there for all companies at once.

Changed the way columns in the mail are created, it now only shows the columns that are used for office/mailbox/office365.

Included a way to exclude more than one OU.